Feature/wcs (#288)
* HOCS-1061
Changing the deployment script for the build pipeline on the WCS branch.

* HOCS-1061
Changing the deployment script for the build pipeline on the WCS branch.

* HOCS-1275
No audit attempts at writing to the queue are visible in the logs. So putting in more debugs and changing the nesting of the try - catches to capture any exception.

* HOCS-1275
No audit attempts at writing to the queue are visible in the logs. So putting in more debugs and changing the nesting of the try - catches to capture any exception.

* HOCS-1275
More logging to find the issue with the stuck queue.

* HOCS-1275
Returning to its previous state before the debugging.

* HOCS-1310 - Allocations report extract is not recording internal stage changes (on/off hold)

Adding ability to allocate user (stage owner) at stage creation time.

* HOCS-1318 : update summary tab

* HOCS-1310 - Allocations report extract is not recording internal stage changes (on/off hold)
recording stage recreation events

* HOCS-1368 - Add WCS reference search to search page

* HOCS-1357 (#280)

Changing the queue settings to ACP system settings at https://github.com/UKHomeOffice/acp-tf-sqs

* HOCS-1370 - Add current stage to case data extract

* HOCS-1360 : task force team upload docs

* changed validation field in field dto from string array to object array

* HOCS-1402 : calculate total amounts

* HOCS-1402 : calculate total amounts, pass back results

* HOCS-1402 : calculate total amounts, use entity list

* Reduce default cache expiration from 10 hours to 1 hour (#284)

Co-authored-by: Paul Rau <52495258+Paulcr-ho@users.noreply.github.com>
Co-authored-by: Hubert Raich <52404016+hubert-raich@users.noreply.github.com>
Co-authored-by: Rviswas <33393268+Rviswas@users.noreply.github.com>
Co-authored-by: Raysonviswas <43213211+Raysonviswas@users.noreply.github.com>