HOCS-2195: fix uuid index and remove redundant index
When creating the somu_item table an incorrect index was created 'idx_somu_item_uuid' that pointed to somu_item(case_uuid). This should have been pointing to somu_item(uuid).

There is also a redundant index in 'idx_somu_item_case_uuid' as a result of 'idx_somu_item_case_uuid_type'. This is because indexes work left to right in subsets. so this indexes case_uuid and case_uuid and type together.