HOCS-2196: merge the create and update unit tests
As a change to design we are now going to have a upsert function that allows for a user to insert or update if the somu type exists. This is reflected in the tests for future development.

Due to getting rid of the update functionality, this also changes the way that we delete, we now have a single delete function instead of previously using the update.