HOCS-2196: fix pr comments
As mentioned by a colleague, it's better to use static data for unit tests. As a result this commit fixes this.

Add the property names and type to the audit payload. 

Change the calls to sentAuditMessage to include the data in the audit data. This allows for it to be sent to hocs-search for indexing on Elastic Search