HOCS-2196: Multiple Contribution Somu Endpoints (#355)
* HOCS-2196 Casework Sumo Item Implementation
This is an intermediate commit so that the changes can be cherry picked onto the MS epic branch.

* HOCS-2196: tests + add isDeleted flag for model

Add tests for the Somu Item model and also add the isDeleted flag on whether the data is null or not. This gets around having the flag in the database which would cause a constraint issue.

* HOCS-2196: add deleted json to dto and tests

Now that we have a deleted flag on the somu_item we bring this into the response object we return through the API. Tests have been added for the DTO object.

* HOCS-2196: add somu item repository test

Add missing tests for the SomuItem repository. These include testing the call to 'findByUuid', 'findAllByCaseUuid' and 'findByCaseUuidAndSomuUuid'.

* HOCS-2196: add audit client tests

This commit adds the missing tests for the audit client with regards to creating, deleting, updating and viewing somu items. Change to the AuditPayload to specify the SomuItem, this is then used within the Audit client for the Audit payload.

* HOCS-2196: add tests for Somu Item Service

This commit adds the necessary tests for the somu item service. Alongside this it also adds a @EqualsAndHashCode linked to the uuid, this will allow for equals between the objects using the uuid which has a unique constraint.

* HOCS-2196: add somu item resource and tests

This commit adds the necessary tests and endpoints to allow for the retrieving of the somu items linked to a case. The following endpoints have been added: 

- Delete:         /item/{somuUuid}
- Update:       /item/{somuUuid}
- Creation:     /case/{caseUuid}/item/{somuTypeUuid}
- Get 1:           /case/{caseUuid}/item/{somuTypeUuid} 
- Get Cases:  /case/{caseUuid}/item

* HOCS-2196 Casework Sumo Item Implementation
Authorisation levels added.
Json Production Format specified on REST calls.

* HOCS-2196: merge the create and update unit tests

As a change to design we are now going to have a upsert function that allows for a user to insert or update if the somu type exists. This is reflected in the tests for future development.

Due to getting rid of the update functionality, this also changes the way that we delete, we now have a single delete function instead of previously using the update.

* HOCS-2196: fix pr comments

As mentioned by a colleague, it's better to use static data for unit tests. As a result this commit fixes this.

Add the property names and type to the audit payload. 

Change the calls to sentAuditMessage to include the data in the audit data. This allows for it to be sent to hocs-search for indexing on Elastic Search

Co-authored-by: Paul Rau <paul.rau1@digital.homeoffice.gov.uk>