HOCS-2240 Use sed instead of variable substitution
Drone 1.0 does a good job of emulating various Bash string operations,
but this doesn't seem to be supported in Drone 0.8. This commit changes
the behaviour to a hacky workaround with sed, which should do the job.

When we migrate to Drone 1.0 we can tidy this up again by reverting this
commit.