HOCS-2235: upsert if not found
When we first load the list, we do a check to see if a case has a particular somu type. If this does not exists the current functionality throws a EntityNotFoundException. This is not ideal as we don't retrieve this on the frontend. A correction to this is to enforce one is created if one isn't found.