HOCS-2274: SomuItems acts as 1-Many
Change to the way that we create new somu items, we no longer have a 1-1 relationship in the SomuItem table to a case and somu type. Each Somu Item is now an individual row in the table. This change is to aid with auditing and being able to better identify what items are added, deleted and updated. Refactored method names to make it clearer what they are doing.