HOCS-2199: Change to audit payloads
At present we are currently trying to map a string to JSON, which results in invalid JSON being received by the audit service. This change encompasses this by having a basic SomuItem class that contains just the SomuTypeUUID for the View. This class is inherited when we do a update, add or delete of an SomuItem to allow for further information being passed.