HOCS-2309: revert change to updateCaseDataValue method (#370)
This change reverts a change that was believed to be required for Multiple Contributions. Previously it was believed we were limited to sending JSON objects from the frontend to the backend. This was not the case and this change caused a regression. This change for MC has been reverted and the corresponding fix has been applied to the front-end that allows for sending of plain text. This therefore fixes the regression.