HOCS-2640: refactor deploy script (#55)
Currently we have a if/else tower that handles the setting of the
`KUBE_TOKEN` environment variable. As we are now passing this
in directly this is no longer needed. Alongside this, we now set the
pipeline to fail if it returns a none 0 status code from a command.
Improve to only have to use the VERSION parameter.