HOCS-2975: change endpoint protection
Currently if we want to update case data, the case has to be 
allocated to the user. As we now allow for WCS cases to be moved 
without allocation. This needs to be changed to accommodate the 
CaseworkTeamUUID data update. To still have some protection, this 
change makes it so the user has to be part of a team that allows 
WRITE permissions.