HOCS-2822: add csv case data updater script (#399)
* HOCS-2822: add csv case data updater script

At present to update cases within the system we have to manually 
create the url requests to be able to update the data or change the 
live data through the frontend. This is easily error prone and 
requires a lot of time to achieve. To improve this mechanism this 
change adds a scripts that takes an absolute csv filepath and 
updates the data based on whats passed in. A new dry run option 
has been added whereby an `update` (case insensitive) second arg 
should be provided to actually send the requests. To make it clear 
what each of the updates are going to do, the script also logs each 
request and the value that it is updating. If the request is successful 
we now also log that. 
As the script is embedded within the `hocs-casework` image, we 
can call the service directly through localhost and use the injected 
SERVER_PORT environment variable.

* HOCS-2822: improve logging format