HOCS-3421: increase hikari connection pool size (#407)
During a Live Service incident we saw that there was occurences
whereby the default connection pool size of 10 were all being used
up. This was causing the pods to crash and fail the health check.
To remediate from this, the default connection pool size has been
increased to 20 while we investigate what the correct number for
the system should be.