HOCS-3013: add dashboard endpoint
At present when the frontend makes a call to load the dashboard,
the `/stage` endpoint is hit. This call returns more information
then is actually required to show the dashboard. As such, this
change adds a `/dashboard` endpoint that returns only the
required information to show the dashboard. 

As the information required is aggregated from a number of
different tables, splitting these out to increase performance was
ideal. This solution therefore queries each of the key pieces of
information separately and zips them together to return
information for each stage/team. As a result of these queries not
returning a entity, we have to run native queries and reflect the
values into a new model Statistic.