HOCS-3645: add correspondent type display name (#420)
* HOCS-3645: add correspondent type display name

Add logic to send the correspondent type display name back with the
correspondents for the People frontend tab.

* HOCS-3645: make BaseCorrespondent abstract

BaseCorrespondent should act as a base and thus should be abstract.

* HOCS-3645: add typeDisplayName to response

Add the typeDisplayName that has been added during the request to the
response. This will then allow the frontend to utilise this.

* HOCS-3645: add all correspondent type call

Add the ability to call the info services method to get all the
correspondent types for use with populating the display name.

* HOCS-3645: add correspondent type name decorator

Add a decorator that will take a list of CorrespondentTypeDto and
Correspondents and populate the `typeDisplayName` if the value is there.
This is then used on the frontend to display the correct name to the
 user.

* HOCS-3645: use the new decorator

User the new CorrespondentTypeNameDecorator within the
CorrespondentService to populate the display name if it is present.

* HOCS-3645: update integration tests

Update integration tests for the new correspondence flow.