HOCS-3848: add isContribution method to inspector
Since the merge in from FoI, there are now SOMU items that can be
non-contributions. For this reason we need to be able to check if an
item is a contribution or not. ReadContext's don't have the ability to
peek at JSON and throws an Exception if the JSON path doesn't exist.
This change checks for a common field across all contributions and runs
a read for it. If an exception is thrown then we know the SOMU Item
isn't a contribution.