HOCS-3848: return FoI contributions (#451)
* HOCS-3848: return contributions for FoI stages

When a stage is at the FOI_APPEALS or FOI_TRIAGE, this changes makes it
so that we process the contributions.

* HOCS-3848: add isContribution method to inspector

Since the merge in from FoI, there are now SOMU items that can be
non-contributions. For this reason we need to be able to check if an
item is a contribution or not. ReadContext's don't have the ability to
peek at JSON and throws an Exception if the JSON path doesn't exist.
This change checks for a common field across all contributions and runs
a read for it. If an exception is thrown then we know the SOMU Item
isn't a contribution.

* HOCS-3848: add filter for contributions

When processing contributions we now filter all of the SOMU items
returned and only return the contributions.