HOCS-3979: contribution processor improvements (#453)
* HOCS-3979: add find in JPA method

Add a JPA method for retrieving the somu items related to a set of case
uuids.

* HOCS-3979: remove retry from audit

There is currently a retry annotation on the audit request, this
annotation does not work on private methods and should be removed as
this code feasibly shouldn't break.

* HOCS-3979: rework contribution processor

Presently, when a workstack is loaded, a call is made for each 'stage'
is part of that workstack to retrieve the contributions. This is causing
 potentially thousands of calls and taking more resources than necessary
This change requests all the contributions for all the stages at once,
leading to only one hit on the database.

A side effect of this is we can revert the auditing code that was
applied previously.