HOCS-3848: rework contributions processor (#454)
* HOCS-3848: add contribution POJO

As we have multiple connotations for what a Contribution is, it is
simpler to hand off the deserialization to Jackson and for us to tell it
 what a contribution can be. This POJO handles this.

* HOCS-3848: rework contribution processor

Previously the contribution processor was heavily influenced by a single
 format for a contribution. This change simplifies this through the use
 of the Jackson library. This change removes the dependency on JSONPath
 which was unnecessary for this work.

* HOCS-3848: correct overdue calculation

The previous overdue calculation was incorrectly showing 18-10-2021 as
due when the date passed in to check again was 19-10-2021. This changes
fixes this by removing the plus 1 day onto the calculation.