HOCS-4045: update log line
Currently, we are printing out all the case uuids that we are
requesting the somu items for in the SomuItemService. This is causing
extremely long logs in production that get split over multiple lines
every time a call to the callee happens. This is a change to the log
line to print out the number of cases that and somu items that have been
 retreived as printing potentially thousands of uuids isn't particulary
 useful.