HOCS-4180 Simplify topic 'findAllByCaseUUID'
This commit replaces the use of active_case in 'findAllByCaseUUID' to provide the original intent of the query as active_case has changed. This query returns all Correspondents by CaseUUID with an additional is_primary flag that haven't been soft deleted and aren't on a Case that has been soft deleted.