HOCS-4182 Expensive calls to GetStageCaseUUIDs This commit replaces the use of active_case to provide the original intent of the query as active_case has changed. This change reduces the execution time from 2 seconds to 0.3ms
This commit also changes the query used by the search to remove an unnecessary table as the field in the where clause was already available on the main table.
This change reduces the execution time from 3 seconds to 1.4 seconds

This commit also introduces an index on stage but used in active_case to reduce query time.
This change reduces the execution time from 450ms  to 1.1ms