HOCS-4182: Move FindAllByCaseUUID to use BasicStage.
This change reduced the query time from 1.3ms to 0.2ms.

More importantly because this change returned a Set of Stage objects for the SAME caseUUID it was returning 'n' copies of the same data blob for the case, this is different from other queries in the repository which return a Set of Stages for different cases. None of the consumers of this endpoint use the data blob and instead (correctly) call the GetCaseData method first meaning we are just wasting bandwidth.