HOCS-4373: rework user permissions
Presently, a lot of our checks revolve around getting a users teams and
then calculating the case types that user is a case admin for. While
this works, if we ever need them teams we have to search for all the
teams for them case types separately. This change offers an 'expanded'
view of a users teams whereby all the team uuids that a user is a member
of, or a case admin of are added together. This will support future
permission development for ensuring users have the correct access to
workstacks and teams.

A side effect of this change is we gain a performance bonus by removing
the case type check this simplifies the queries that used it. From
`12329.683ms` to `731.688ms` on a production dataset.