HOCS-3953: rework notify SQS integration
This change removes the camel processor and usage from the SQS
integration. This involved replacing the camel processor with the direct
Amazon SQS client. This simplified the code flow through this and
enabled real integration tests. We now also ensure that the correct
headers are passed through.

As the properties have been generalised, this caused changes to be
required with the deployment yaml and the corresponding application
.properties files.

A side effect of this change is all non-mock tests now require the
`local` profile to instantiate and run.