HOCS-3953: rework SNS integration
This change removes the camel processor and usage from the SNS
integration. This involved replacing the camel processor with the direct
Amazon SNS client. This simplified the code flow through this and
enabled real integration tests. We now also ensure that the correct
headers are passed through.

As the properties have been generalised, this caused changes to be
required with the deployment yaml and the corresponding application
.properties files.