HOCS-4724: send final stage completed audit
When a case is at the end of a workflow, when completing a case we don't
send a STAGE_COMPLETED audit line for that stage. This change triggers
this audit event when a case is completed with an active stage.

(cherry picked from commit c5140058d8179178bcae463804cda03a66155ba1)