Fix dependancy issues (#630)
* Fix Hibernate-Types Dependency

The latest 2.15.1 version of the `hibernate-types` requires the 'Range'
class of Guava (which it does not provide). As we have Guava in our test
 dependencies this change passes current tests. Changing this to a
 general `implementation` gives runtime access to the library.

* Use more suitable hibernate-types library

We use `hibernate-types-52` which is suitable for Hibernate
`5.4, 5.3, 5.2`. The hibernate library is currently 5.6.7 which doesn't
have a hibernate-types library. This change upgrades to the closest
match for the time being.

* Remove hibernate types banner