HOCS-5240: allow return of deleted correspondents
Change to the GET correspondent endpoint to support an Optional request
param for getting all correspondents including deleted ones. As this is
optional the existing behaviour is default and anything requiring can
opt in.

This has caused a re-write and simplification of the Correspondents
objects and JPA implementation. It now defaults to using the
Correspondent and on the endpoint required it retrieves the case data
and mark the correct correspondent with the primary flag.