HOCS-5278: change field restriction behaviour
This change causes fields to be removed from the case data on restricted
cases by exemption rather than an inclusion policy. This is done
through the `RestrictedField[Service/Repository]` instead of  behaviour
added previously to the UserPermissionService that communicated with
`hocs-info-service`.

As we are handling changes to collections whereby the objects within a
mutable - this change removes the need for helper static classes.