HOCS-5354: remove CASE_TYPE_MARKER
Previously there was a necessary case type marker that was added to the
case data blob during priority calculation that allowed for exemption
dates to be returned. As the whole StageWithCaseData object is passed
now, the case type is available in the resultant call. Therefore, this
marker is no longer required and has been removed.