HOCS-5569: add services for details
Add the services and repository surrounding loading of the case
details (unallocated view) from an group/ stage configuration. As the
`group` way is what we are envisaging to move towards, this is an early
retrieval and exit. As current CS case types run on a stage based
system, the fallback is to look at the stage configuration folder and
return a dataset that is like the current version.