HOCS-5690: add case data tag table
Add the case data tag table, this contains the case uuid and tag as the
primary composite key. Alongside this we store a created_at and
deleted_on timestamp to enable any potential future investigations. The
change also provides an index based on deleted_on being null to allow
for quick lookups.