HOCS-5920: rework tagging functionality
There are currently performance issues with the tagging functionality
whereby the use of native queries and JPA hibernate is causing queries
per stage for tags.

While this functionality is being rewritten, this change falls back to
using a 'somu'-esque loading mechanism to only query all stage tags in
one query and then zip into the stages.