HOCS-6264: correct stage override date setting
There was a regression in functionality with the implementation of FOI
Dispatch. This resulted in DCU cases going through the same code path
with an override being ignored. This causes issues with deadlines, and
therefore business SLAs being met. This change corrects this code path
so that FOI extensions do not impact overrides.